requirements.txt
# --- Web framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5

# --- Cryptography ---
cryptography==44.0.0
PyNaCl==1.5.0

# --- HTTP client ---
httpx==0.28.1
requests==2.32.3

# --- Data ---
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# --- HashiCorp Vault client ---
hvac==2.3.0

# --- Utilities ---
base58==2.1.1
